feat(repositories): BaseEntity now uses Long id, updated Repositories accordingly

+++ src/main/java/dev/steelbookdb/steelbookapi/localization/CountryRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.localization;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountryRepository extends JpaRepository<Country, Integer> {
+public interface CountryRepository extends JpaRepository<Country, Long> {
 
 }
