+++ src/main/java/dev/steelbookdb/steelbookapi/movie/MovieRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.movie;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Integer> {
+public interface MovieRepository extends JpaRepository<Movie, Long> {
 
 }
