+++ src/main/java/dev/steelbookdb/steelbookapi/movie/MovieTranslationRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.movie;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieTranslationRepository extends JpaRepository<MovieTranslation, Integer> {
+public interface MovieTranslationRepository extends JpaRepository<MovieTranslation, Long> {
 
 }
