+++ src/main/java/dev/steelbookdb/steelbookapi/steelbook/AudioFormatRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.steelbook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AudioFormatRepository extends JpaRepository<AudioFormat, Integer> {
+public interface AudioFormatRepository extends JpaRepository<AudioFormat, Long> {
 
 }
