+++ src/main/java/dev/steelbookdb/steelbookapi/steelbook/DiskRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.steelbook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiskRepository extends JpaRepository<Disk, Integer> {
+public interface DiskRepository extends JpaRepository<Disk, Long> {
 
 }
