+++ src/main/java/dev/steelbookdb/steelbookapi/steelbook/EditorRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.steelbook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EditorRepository extends JpaRepository<Editor, Integer> {
+public interface EditorRepository extends JpaRepository<Editor, Long> {
 
 }
