+++ src/main/java/dev/steelbookdb/steelbookapi/steelbook/RetailerRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.steelbook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RetailerRepository extends JpaRepository<Retailer, Integer> {
+public interface RetailerRepository extends JpaRepository<Retailer, Long> {
 
 }
