+++ src/main/java/dev/steelbookdb/steelbookapi/steelbook/SteelbookRepository.java
@@ -2,6 +2,6 @@ package dev.steelbookdb.steelbookapi.steelbook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SteelbookRepository extends JpaRepository<Steelbook, Integer> {
+public interface SteelbookRepository extends JpaRepository<Steelbook, Long> {
 
 }
